web_demo/requirements.txt
matplotlib==3.6.3
Pillow==9.4.0
streamlit==1.17.0
onnx==1.14.0
onnxruntime==1.14.1
torchvision==0.14.1
